app/requirements.txt
Flask-Markdown==0.3
Flask-MySQLdb==1.0.1
Flask==1.1.2
itsdangerous==2.0.1
Pygments==2.11.2
bs4==0.0.1
mysql-connector-python==8.0.28
py-gfm==1.0.2

